Experimental partial class
